Move guestbook tests into its own file
