+++ compiler/tests/evm_contracts.rs
@@ -303,29 +303,6 @@ fn address_bytes10_map() {
 }
 
 #[test]
-fn guest_book() {
-    with_executor(&|mut executor| {
-        let mut harness = deploy_contract(&mut executor, "demos/guest_book.fe", "GuestBook", &[]);
-
-        let sender = address_token("1234000000000000000000000000000000005678");
-        let bytes = bytes_token(
-            iter::repeat("ten bytes.")
-                .take(10)
-                .collect::<String>()
-                .as_str(),
-        );
-
-        harness.caller = sender.clone().to_address().unwrap();
-
-        harness.test_function(&mut executor, "sign", &[bytes.clone()], None);
-
-        harness.test_function(&mut executor, "get_msg", &[sender], Some(&bytes));
-
-        harness.events_emitted(executor, &[("Signed", &[bytes])]);
-    })
-}
-
-#[test]
 fn return_builtin_attributes() {
     let gas_price = 123;
     let origin = address_token("0000000000000000000000000000000000000001");
